till now updated

+++ OnlineCourierManagementApplication/src/main/java/com/capg/ocma/repository/IComplaintDao.java
@@ -4,10 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.capg.ocma.entities.Complaint;
-import com.capg.ocma.entities.Courier;
+
 
 @Repository
-public interface IComplaintDao extends JpaRepository<Complaint, Long> {
+public interface IComplaintDao extends JpaRepository<Complaint, Integer> {
